LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/multicol-fragmented-anchor-expected.txt

content

FAIL An element in a fragmented div should be able to be selected as an anchor node. assert_equals: expected 20 but got 10
